+++ prisma/schema.prisma
@@ -15,14 +15,15 @@ datasource db {
 }
 
 model User {
-  id          Int      @id @default(autoincrement())
-  email       String   @unique
+  id          Int        @id @default(autoincrement())
+  email       String     @unique
   external_id String
   createdAt   DateTime   @default(now())
   updatedAt   DateTime   @updatedAt
 }
 
 model Organization {
-  id      Int      @id @default(autoincrement())
-  name    String
+  id          Int      @id @default(autoincrement())
+  external_id String   @unique
+  name        String
 }
